applications: nrf_desktop: Introduce HID keymap utility
Change introduces utility that can be used to map app-specific key IDs
to HID usage IDs. Change also integrates the utility in HID state module
and aligns application configurations.

Jira: NCSDK-33877
