+++ applications/nrf_desktop/configuration/nrf52840dk_nrf52840/prj_keyboard.conf
@@ -22,7 +22,7 @@ CONFIG_DESKTOP_HID_BOOT_INTERFACE_KEYBOARD=y
 
 # Application supports 2 HID subscribers: HID over GATT (BLE) and a single USB HID instance
 CONFIG_DESKTOP_HID_STATE_SUBSCRIBER_COUNT=2
-CONFIG_DESKTOP_HID_STATE_HID_KEYMAP_DEF_PATH="hid_keymap_def_keyboard.h"
+CONFIG_DESKTOP_HID_KEYMAP_DEF_PATH="hid_keymap_def_keyboard.h"
 CONFIG_DESKTOP_HID_STATE_HID_KEYBOARD_LEDS_DEF_PATH="hid_keyboard_leds_def_keyboard.h"
 CONFIG_DESKTOP_HID_REPORT_EXPIRATION=10000
 CONFIG_DESKTOP_HID_EVENT_QUEUE_SIZE=60
