+++ applications/nrf_desktop/configuration/nrf54l15dk_nrf54l10_cpuapp/prj_keyboard.conf
@@ -18,7 +18,7 @@ CONFIG_DESKTOP_ROLE_HID_PERIPHERAL=y
 CONFIG_DESKTOP_PERIPHERAL_TYPE_KEYBOARD=y
 CONFIG_DESKTOP_DEVICE_PID=0x52DD
 CONFIG_DESKTOP_HID_BOOT_INTERFACE_KEYBOARD=y
-CONFIG_DESKTOP_HID_STATE_HID_KEYMAP_DEF_PATH="hid_keymap_def_keyboard.h"
+CONFIG_DESKTOP_HID_KEYMAP_DEF_PATH="hid_keymap_def_keyboard.h"
 CONFIG_DESKTOP_HID_STATE_HID_KEYBOARD_LEDS_DEF_PATH="hid_keyboard_leds_def_keyboard.h"
 CONFIG_DESKTOP_HID_REPORT_EXPIRATION=10000
 CONFIG_DESKTOP_HID_EVENT_QUEUE_SIZE=60
